separate kyma-metrics with deletion
