+++ tests/e2e_test/Makefile
@@ -38,6 +38,10 @@ test: test-watcher-e2e test-kyma-deletion-e2e test-status-propagation-e2e
 test-kyma-deletion-e2e: ## Runs the Kyma Deletion E2E Test
 	go test --tags=deletion_e2e -ginkgo.vv
 
+.PHONY: kyma-metrics
+kyma-metrics: ## Runs the Kyma Metrics E2E Test
+	go test --tags=kyma_metrics -ginkgo.vv
+
 .PHONY: test-watcher-e2e
 test-watcher-e2e: ## Runs the Watcher E2E Test
 	go test --tags=watcher_e2e -ginkgo.vv
