+++ tests/e2e_test/kyma_deletion_test.go
@@ -119,66 +119,3 @@ var _ = Describe("KCP Kyma CR should be deleted successfully when SKR cluster ge
 				Should(Succeed())
 		})
 	})
-
-var _ = Describe("When KCP Kyma CR deleted, Kyma Status Metric should also be deleted",
-	Ordered, func() {
-		channel := "regular"
-		kyma := testutils.NewKymaForE2E("kyma-sample", "kcp-system", channel)
-		remoteNamespace := "kyma-system"
-		GinkgoWriter.Printf("kyma before create %v\n", kyma)
-
-		BeforeAll(func() {
-			// make sure we can list Kymas to ensure CRDs have been installed
-			err := controlPlaneClient.List(ctx, &v1beta2.KymaList{})
-			Expect(meta.IsNoMatchError(err)).To(BeFalse())
-		})
-
-		It("Should create empty Kyma CR", func() {
-			Eventually(CreateKymaSecret, timeout, interval).
-				WithContext(ctx).
-				WithArguments(kyma.GetName(), kyma.GetNamespace(), controlPlaneClient).
-				Should(Succeed())
-			Eventually(controlPlaneClient.Create, timeout, interval).
-				WithContext(ctx).
-				WithArguments(kyma).
-				Should(Succeed())
-			By("verifying kyma is ready")
-			Eventually(CheckKymaIsInState, readyTimeout, interval).
-				WithContext(ctx).
-				WithArguments(kyma.GetName(), kyma.GetNamespace(), controlPlaneClient, v1beta2.StateReady).
-				Should(Succeed())
-			By("verifying remote kyma is ready")
-			Eventually(CheckRemoteKymaCR, readyTimeout, interval).
-				WithContext(ctx).
-				WithArguments(remoteNamespace, []v1beta2.Module{}, runtimeClient, v1beta2.StateReady).
-				Should(Succeed())
-		})
-
-		It("Kyma reconciliation should remove metric when Kyma CR deleted ", func() {
-			By("getting the current kyma Ready state metric count")
-			kymaStateReadyCount, err := GetKymaStateMetricCount(kyma.GetName(), "Ready")
-			Expect(err).Should(Not(HaveOccurred()))
-			GinkgoWriter.Printf("Kyma State Ready Metric count before CR deletion: %d", kymaStateReadyCount)
-			Expect(kymaStateReadyCount).Should(Equal(1))
-
-			By("deleting KCP Kyma")
-			Eventually(controlPlaneClient.Delete, readyTimeout, interval).
-				WithContext(ctx).
-				WithArguments(kyma).
-				Should(Succeed())
-
-			By("waiting for Kyma CR to be removed")
-			Eventually(CheckKCPKymaCRDeleted, 2*time.Minute, interval).
-				WithContext(ctx).
-				WithArguments(kyma.GetName(), kyma.GetNamespace(), controlPlaneClient).
-				Should(Succeed())
-
-			By("should decrease the metric count")
-			for _, state := range []string{"Deleting", "Warning", "Ready", "Processing", "Error"} {
-				count, err := GetKymaStateMetricCount(kyma.GetName(), state)
-				Expect(err).Should(Not(HaveOccurred()))
-				GinkgoWriter.Printf("Kyma %s State Metric count after CR deletion: %d", state, count)
-				Expect(count).Should(Equal(0))
-			}
-		})
-	})
